refactor(event_handler): 提取 MemoryReflection 模块
- 创建 core/event_handler_modules/memory_reflection.py (375 行)
- 将记忆反思和存储逻辑提取到 MemoryReflection 类
- 包含完整的总结触发、待处理失败总结重试、后台任务管理等逻辑
- 包含3个核心方法：
  - handle_memory_reflection: 主入口，检查是否需要总结
  - _storage_task: 后台存储任务
  - _record_pending_summary: 记录失败总结供重试
- EventHandler 中的 handle_memory_reflection 简化为委托调用
- 更新测试以适配新的模块结构
- event_handler.py 从 572 行减少到 142 行 (-430 行, -75.2%)
- 所有 44 个 event_handler 测试通过
