+++ core/event_handler_modules/__init__.py
@@ -4,6 +4,7 @@ EventHandler 模块化子模块
 
 from .group_capture import GroupCapture
 from .memory_recall import MemoryRecall
+from .memory_reflection import MemoryReflection
 from .message_utils import MessageUtils
 
-__all__ = ["MessageUtils", "GroupCapture", "MemoryRecall"]
+__all__ = ["MessageUtils", "GroupCapture", "MemoryRecall", "MemoryReflection"]
